feat: update beans-logging version and add async HTTP error logging functions
