+++ requirements.txt
@@ -1,2 +1,2 @@
 fastapi>=0.99.1,<1.0.0
-beans-logging>=10.0.0,<11.0.0
+beans-logging>=10.0.1,<11.0.0
