+++ src/beans_logging_fastapi/__init__.py
@@ -3,10 +3,13 @@ from beans_logging import logger
 from .__version__ import __version__
 from .config import LoggerConfigPM
 from ._core import add_logger
+from .http_error import async_log_http_error, log_http_error
 
 __all__ = [
     "__version__",
     "logger",
     "add_logger",
     "LoggerConfigPM",
+    "async_log_http_error",
+    "log_http_error",
 ]
